Fix for non-installable package due to '/lib already owned by filesystem’ error
Under Arch Linux and similiar, /lib is always symlink to /usr/lib. Packages trying to put anything in /lib directory subtree fails to install as pacman considers it as an attempt to overwrite the /lib symlink with a directory and refuses to continue. The confusing thing is that this isn’t reported as a warning or error in any way by makepkg, but packages aren’t installable.